+++ gakumas-tools/components/Simulator/Simulator.js
@@ -371,7 +371,7 @@ export default function Simulator() {
             href="https://github.com/surisuririsu/gakumas-tools/blob/master/gakumas-tools/simulator/CHANGELOG.md"
             target="_blank"
           >
-            {t("lastUpdated")}: 2026-04-29
+            {t("lastUpdated")}: 2026-05-01
           </a>
         </div>
       </div>